+++ examples/di/ml_engineer_with_tools.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from metagpt.roles.mi.ml_engineer import MLEngineer
+from metagpt.roles.di.ml_engineer import MLEngineer
 
 
 async def main(requirement: str):
