+++ examples/di/rm_image_background.py
@@ -1,11 +1,11 @@
 import asyncio
 
-from metagpt.roles.mi.interpreter import Interpreter
+from metagpt.roles.di.data_interpreter import DataInterpreter
 
 
 async def main(requirement: str = ""):
-    mi = Interpreter(use_tools=False)
-    await mi.run(requirement)
+    di = DataInterpreter(use_tools=False)
+    await di.run(requirement)
 
 
 if __name__ == "__main__":
