+++ metagpt/actions/di/debug_code.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from metagpt.actions.mi.write_analysis_code import BaseWriteAnalysisCode
+from metagpt.actions.di.write_analysis_code import BaseWriteAnalysisCode
 from metagpt.logs import logger
 from metagpt.schema import Message
 from metagpt.utils.common import create_func_call_config
