+++ tests/metagpt/actions/di/test_ask_review.py
@@ -1,6 +1,6 @@
 import pytest
 
-from metagpt.actions.mi.ask_review import AskReview
+from metagpt.actions.di.ask_review import AskReview
 
 
 @pytest.mark.asyncio
