+++ tests/metagpt/actions/di/test_debug_code.py
@@ -5,7 +5,7 @@
 
 import pytest
 
-from metagpt.actions.mi.debug_code import DebugCode
+from metagpt.actions.di.debug_code import DebugCode
 from metagpt.schema import Message
 
 ErrorStr = """Tested passed:
