+++ tests/metagpt/actions/di/test_write_plan.py
@@ -1,6 +1,6 @@
 import pytest
 
-from metagpt.actions.mi.write_plan import (
+from metagpt.actions.di.write_plan import (
     Plan,
     Task,
     WritePlan,
