+++ tests/metagpt/utils/test_save_code.py
@@ -6,7 +6,7 @@
 import nbformat
 import pytest
 
-from metagpt.actions.mi.execute_nb_code import ExecuteNbCode
+from metagpt.actions.di.execute_nb_code import ExecuteNbCode
 from metagpt.utils.common import read_json_file
 from metagpt.utils.save_code import DATA_PATH, save_code_file
